Improve build.gradle to scan for version overrides with no config, and output debug std out. Wrapped more 1.20.4 methods such as client packet register and send methods, and buffer operations. Fixed effect status to be more compatible.

+++ src/main/java/com/owlmaddie/goals/LeadPlayerGoal.java
@@ -145,7 +145,7 @@ public class LeadPlayerGoal extends PlayerBaseGoal {
             ServerWorld serverWorld = (ServerWorld) this.entity.getWorld();
 
             // Pass the angle using the "speed" argument, with deltaX, deltaY, deltaZ set to 0
-            LeadParticleEffect effect = new LeadParticleEffect(angle);
+            LeadParticleEffect effect = new LeadParticleEffect((float)angle);
             serverWorld.spawnParticles(effect, position.x, position.y + 0.05, position.z, 1, 0, 0, 0, 0);
         }
     }
